langgraph/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.36
langchain-core==0.3.13
google-generativeai==0.7.2
